.history/requirements_20251003014046.txt
# Core dependencies
nltk>=3.8
scikit-learn>=1.3.0
numpy>=1.24.0
networkx>=3.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Evaluation
rouge-score>=0.1.2

# AI/ML
google-generativeai>=0.3.0

# Optional: For better performance
# sentence-transformers>=2.2.0  # Uncomment for advanced embeddings
# spacy>=3.6.0  # Uncomment for advanced NLP
